merge radio_server.py code into metamod PluginExit
simplifies packet logic, improves performance, and reduces latency. No more reading/writing files to send packets between radio server and the game. The plugin connects to the server directly and does packet processing in a secondary thread. The main thread only needs to send the network messages to each player, which it appears to be doing at least 1000x faster than angelscript could.

Linux version is crashing when updated in metamod, and the TCP connection seems touchy and fails a lot a first.